+++ Providers/DBPlanItem.swift
@@ -12,8 +12,8 @@ public class DBPlanItem: DBSyncable {
     
     dynamic var inventoryOpt: DBInventory? = DBInventory()
     dynamic var productOpt: Product? = Product() // this should be quantifiable product but we don't use plan items now so we let it like this for now
-    public dynamic var quantity: Int = 0
-    public dynamic var quantityDelta: Int = 0
+    public dynamic var quantity: Float = 0
+    public dynamic var quantityDelta: Float = 0
     
     public dynamic var key: String = ""
